Fix doubling/undoubling for prefixes
This corrects an issue where doubling/undoubling would always occur at
the end of a stem, even if a prefix is used. Fixing this changes the
analysis of two forms in the Brown corpus:
uncontrollable  (un)+ CONTROL +(able) | uncontrollable  UNCONTROLL* +(able)
uncontrolled    (un)+ CONTROL +(ed)   | uncontrolled    UNCONTROLL* +(ed)

Previously, applying (un)+ to the base "control" would lead to the string
"uncontroll", instead of "unccontrol". This error actually leads to a
correct analysis of "uncontrollable" and "uncontrolled" even though the
system should not be able to identify that relationship.