Rename base inference params and allow recomputing inferred bases every iteration
Recomputing the inferred bases every iteration ensures that bases are not
inferred based on stale information. Without this recomputation, a base
suggested by an earlier iteration which may no longer be needed (because
of new transforms learned) will be used as evidence. This can result in
cases where a new base is inferred to only derive a single form, because
the other forms that suggested the base should exist have already been
derived by other means. On the development Brown wordlist, this very
slightly hurts recall but boosts precision, maintaining an equivalent
F1 score.